Reorganizado estrutura de arquivos do projeto
- move `index.js` para `src/js/index.js`
- move `style.css` para `src/css/themes.css`
- move `phrases.js` para `src/js/data/phrases.js`
- adiciona `favicon.ico` na pasta `public/`
- ajusta caminhos no `index.html` conforme nova estrutura
- remove `root` do `vite.config.js` para usar `index.html` na raiz
