+++ index.html
@@ -16,7 +16,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="icon" type="image/x-icon" href="images/favicon.ico">
-  <link rel="stylesheet" href="/public/styles/themes.css">
+  <link rel="stylesheet" href="src/css/themes.css">
   <title>Binário em Palavras</title>
   
   <style>
@@ -41,7 +41,7 @@
   })();
   </script>
 
-  <script defer type="module" src="index.js"></script>
+  <script defer src="src/js/index.js" type="module"></script>
 </head>
 
 <body class="bg-custom-page p-3">
